fix: renamings etc
