+++ sample-operators/webpage/src/main/java/io/javaoperatorsdk/operator/sample/WebPageOperator.java
@@ -26,8 +26,8 @@ public class WebPageOperator {
     Config config = new ConfigBuilder().withNamespace(null).build();
     KubernetesClient client = new DefaultKubernetesClient(config);
     Operator operator = new Operator(client, DefaultConfigurationService.instance());
+    operator.register(new WebPageReconcilerDependentResources(client));
     operator.register(new WebPageReconciler(client));
-    operator.register(new WebPageReconcilerLowLevel(client));
     operator.installShutdownHook();
     operator.start();
